bot/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
requests==2.32.3
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
qrcode==7.4.2
Pillow==10.4.0

